requirements.txt
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
customtkinter>=5.2.0
mysql-connector-python>=8.0.33
python-dotenv>=1.0.0
pandas>=2.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0 